move page to /start
