+++ packages/client/src/pages/StartPage/index.ts
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export const StartPage = lazy(() => import('./StartPage'));
